# pkg/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for a login API.",
        "title": "Swagger Example API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/login": {
            "post": {
                "description": "this is a example to login in the site",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "login example",
                "parameters": [
                    {
                        "description": "Login request",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.LoginResponse"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "this is the documentation abaut register a new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "register example",
                "parameters": [
                    {
                        "description": "register request",
                        "name": "register",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.RegisterResponse"
                        }
                    }
                }
            }
        },
        "/token": {
            "post": {
                "description": "this is the documentation abaut token to refresh the token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "token auth",
                "parameters": [
                    {
                        "description": "refreshAccessTokenRequest",
                        "name": "refresh",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.RefreshAccessTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.RefreshAccessTokenResponse"
                        }
                    }
                }
            }
        },
        "/user/password": {
            "put": {
                "description": "actualiza la contraseña de algun usuario",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "update a password",
                "parameters": [
                    {
                        "description": "update password request",
                        "name": "Password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.UpdatePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/upgrade": {
            "put": {
                "description": "hace que un usuario se vuelva un super usuario",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "upgrade to a super user",
                "parameters": [
                    {
                        "description": "update password request",
                        "name": "Password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.UpgradeToSuperUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "description": "obtener un usuario por su id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "get user by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/users.GetUserResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "elimina un usuario segun su id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "delete a user by id",
                "parameters": [
                    {
                        "description": "delete user request",
                        "name": "Password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.DeleteUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "description": "obtienes varios usuarios en un rango dado por el pageID and pageSize",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "get many users",
                "parameters": [
                    {
                        "description": "list user  request",
                        "name": "listUser",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.ListUserRequest"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "auth.LoginRequest": {
            "type": "object",
            "required": [
                "Username",
                "password"
            ],
            "properties": {
                "Username": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "auth.LoginResponse": {
            "type": "object",
            "properties": {
                "Email": {
                    "type": "string"
                },
                "RefreshToken": {
                    "type": "string"
                },
                "RefreshToken_expire_at": {
                    "type": "string"
                },
                "Username": {
                    "type": "string"
                },
                "accessToken": {
                    "type": "string"
                },
                "accessToken_expire_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "municipio": {
                    "type": "string"
                },
                "provincia": {
                    "type": "string"
                },
                "session_id": {
                    "type": "string"
                }
            }
        },
        "auth.RefreshAccessTokenRequest": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "auth.RefreshAccessTokenResponse": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "accessToken_expire_at": {
                    "type": "string"
                }
            }
        },
        "auth.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "municipio",
                "password",
                "password2",
                "provincia",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "municipio": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "password2": {
                    "type": "string"
                },
                "provincia": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "auth.RegisterResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "municip": {
                    "type": "string"
                },
                "provincia": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "users.DeleteUserRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "users.GetUserResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "municipio": {
                    "type": "string"
                },
                "provincia": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "users.ListUserRequest": {
            "type": "object",
            "required": [
                "page_id",
                "page_size"
            ],
            "properties": {
                "page_id": {
                    "type": "integer",
                    "minimum": 1
                },
                "page_size": {
                    "type": "integer",
                    "maximum": 10,
                    "minimum": 5
                }
            }
        },
        "users.UpdatePasswordRequest": {
            "type": "object",
            "required": [
                "id",
                "password",
                "password2"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                },
                "password2": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
        "users.UpgradeToSuperUser": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        }
    }
}